// Angular Frontend/src/app/model/hotelowner.ts
export interface HotelOwner {

    hotelOwnerName: string;
    username: string;
    password: string;
    email: string;
    gender: string;
    address: string;
    hotelName: string;
    location: string;
    hasDining: boolean;
    hasParking: boolean;
    hasFreeWiFi: boolean;
    hasRoomService: boolean;
    hasSwimmingPool: boolean;
    hasFitnessCenter: boolean;

}